Restructure directories to be consistent to StatisticsViews
