+++ app/src/components/ExplorationViews/MethodView/MethodView.jsx
@@ -3,7 +3,7 @@ import SyntaxHighlighter from 'react-syntax-highlighter';
 import { docco } from 'react-syntax-highlighter/dist/cjs/styles/hljs';
 import PropTypes from 'prop-types';
 import React, { useEffect, useState } from 'react';
-import { getMethodInfo, getMethodText } from '../../utils/apiHandler';
+import { getMethodInfo, getMethodText } from '../../../utils/apiHandler';
 import './MethodView.css';
 
 const MethodView = ({ currentClass, site, currentMethod }) => {
